move test files, update require paths
